requirements.txt
numpy
opencv-python
pillow
PythonTurtle
jupyterlab
notebook
jupyterthemes